Initial commit
It contains three network models describing political preferences: Barabasi-Albert, Watts-Strogatz, Erdos-Renyi.
To show the evolution of each social network, `update_opinion` function is applied to all nodes in graph.
On each corner of the plot, there is point indicating political affair of the each node, for better visibility of the changes direction in the network.